+++ src/app/service/bounty/create/PublishBounty.ts
@@ -2,7 +2,7 @@ import constants from '../../constants/constants';
 import mongo, { Db, UpdateWriteOpResult } from 'mongodb';
 import BountyUtils from '../../../utils/BountyUtils';
 import { GuildMember, Message, MessageEmbedOptions, TextChannel } from 'discord.js';
-import dbInstance from '../../../utils/db';
+import dbInstance from '../../../utils/dbUtils';
 import channelIDs from '../../constants/channelIds';
 import ServiceUtils from '../../../utils/ServiceUtils';
 import envUrls from '../../constants/envUrls';