+++ src/app/utils/POAPUtils.ts
@@ -1,9 +1,10 @@
 import { GuildChannel, GuildMember, MessageAttachment } from 'discord.js';
-import { Collection as MongoCollection, Cursor, Db, UpdateWriteOpResult } from 'mongodb';
+import { Collection, Collection as MongoCollection, Cursor, Db, UpdateWriteOpResult } from 'mongodb';
 import constants from '../service/constants/constants';
 import { POAPParticipant } from '../types/poap/POAPParticipant';
 import axios from 'axios';
 import ValidationError from '../errors/ValidationError';
+import { POAPAdmin } from '../types/poap/POAPAdmin';
 
 export type POAPFileParticipant = {
 	id: string,
@@ -71,7 +72,7 @@ const POAPUtils = {
 		}
 		for (let i = 0; i < listOfParticipants.length; i++) {
 			try {
-				guildMember.guild.members.fetch(listOfParticipants[i].id)
+				await guildMember.guild.members.fetch(listOfParticipants[i].id)
 					.then(async (participantMember: GuildMember) => {
 						await participantMember.send({ content: `Thank you for participating in BanklessDAO! Here is your POAP: ${listOfPOAPLinks[i]}` }).catch((e) => {
 							console.log(`failed trying to send POAP to: ${listOfParticipants[i].id}, userTag: ${listOfParticipants[i].tag}, link: ${listOfPOAPLinks[i]}`);
@@ -86,11 +87,11 @@ const POAPUtils = {
 							console.error(e);
 						});
 					});
-				console.log(`Links sent to ${listOfParticipants.length} participants.`);
 			} catch (e) {
 				console.log('user might have been banned');
 			}
 		}
+		console.log(`Links sent to ${listOfParticipants.length} participants.`);
 	},
 
 	async validateEvent(guildMember: GuildMember, event?: string): Promise<any> {
@@ -108,6 +109,30 @@ const POAPUtils = {
 			});
 			throw new ValidationError('Please try another event.');
 		}
+	},
+	
+	async validateUserAccess(guildMember: GuildMember, db: Db): Promise<any> {
+		const poapAdminsDb: Collection = await db.collection(constants.DB_COLLECTION_POAP_ADMINS);
+		const userResult: POAPAdmin = await poapAdminsDb.findOne({
+			objectType: 'USER',
+			discordObjectId: guildMember.user.id,
+			discordServerId: guildMember.guild.id,
+		});
+		if (userResult != null) {
+			// user has access
+			return;
+		}
+		const rolesCursor: Cursor<POAPAdmin> = await poapAdminsDb.find({
+			objectType: 'ROLE',
+			discordServerId: guildMember.guild.id,
+		});
+		for await (const poapRole of rolesCursor) {
+			if (guildMember.roles.cache.some(role => role.id === poapRole.discordObjectId)) {
+				// role has access
+				return;
+			}
+		}
+		throw new ValidationError('You are not configured to use this command. Please reach out to discord owner.');
 	},
 };